Add test for items count
